requirement.txt
django-environ==0.10.0
django-authtools
djangorestframework